requirements.txt
addict==2.4.0
aenum==3.1.11
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
carla==0.9.14
click==8.1.3
colorama==0.4.6
comm==0.1.3
ConfigArgParse==1.5.3
contourpy==1.0.7
cycler==0.11.0
dash==2.9.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
fastjsonschema==2.16.3
Flask==2.2.3
fonttools==4.39.3
future==0.18.3
h5py==3.8.0
imageio==2.27.0
importlib-metadata==6.1.0
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==8.12.0
ipywidgets==8.0.6
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
llvmlite==0.41.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
nbformat==5.5.0
nest-asyncio==1.5.6
networkx==3.1
numba==0.58.0
numpy==1.24.2
open3d==0.16.0
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.2.0
plotly==5.14.0
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pygame==2.3.0
Pygments==2.14.0
pyparsing==3.0.9
pyquaternion==0.9.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
scikit-learn==1.0
scipy==1.10.1
shapely==2.0.2
six==1.16.0
stack-data==0.6.3
tenacity==8.2.3
threadpoolctl==3.2.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
typing_extensions==4.9.0
wcwidth==0.2.13
Werkzeug==3.0.1
widgetsnbextension==4.0.9
zipp==3.17.0
